upload-service/requirements.txt
Flask==2.2.2
Flask-Cors==3.0.10
Werkzeug==2.2.2
python-dotenv